# gen/models/subscribe.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// Subscribe subscribe
//
// swagger:model Subscribe
type Subscribe struct {

	// item id
	// Required: true
	ItemID *int64 `json:"item_id"`

	// status
	Status *bool `json:"status,omitempty"`

	// subs price
	// Required: true
	SubsPrice *float64 `json:"subs_price"`
}

// Validate validates this subscribe
func (m *Subscribe) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateItemID(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateSubsPrice(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Subscribe) validateItemID(formats strfmt.Registry) error {

	if err := validate.Required("item_id", "body", m.ItemID); err != nil {
		return err
	}

	return nil
}

func (m *Subscribe) validateSubsPrice(formats strfmt.Registry) error {

	if err := validate.Required("subs_price", "body", m.SubsPrice); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this subscribe based on context it is used
func (m *Subscribe) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *Subscribe) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Subscribe) UnmarshalBinary(b []byte) error {
	var res Subscribe
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}