# gen/models/subscription_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// SubscriptionResponse subscription response
//
// swagger:model SubscriptionResponse
type SubscriptionResponse struct {

	// message
	Message string `json:"Message,omitempty"`

	// success
	Success bool `json:"success,omitempty"`
}

// Validate validates this subscription response
func (m *SubscriptionResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this subscription response based on context it is used
func (m *SubscriptionResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *SubscriptionResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *SubscriptionResponse) UnmarshalBinary(b []byte) error {
	var res SubscriptionResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}