# swagger.yml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "ustore"
host: "localhost:8080"
basePath: "/v1"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /item/subscribe:
    post:
      description: 'To subscribe an available item'
      operationId: "subscribe"
      tags:
        - "item"
      security:
        - Bearer: [ ]
      parameters:
        - in: 'body'
          name: 'subscribe'
          required: true
          description: "subscription model"
          schema:
            $ref: "#/definitions/Subscribe"
      responses:
        200:
          description: Successful Subscription
          schema:
            $ref: '#/definitions/SubscriptionResponse'
        400:
          description: Bad Request
        404:
          schema:
            type: string
          description: Item not found
        500:
          schema:
            type: string
          description: Server error
definitions:
...
  Subscribe:
    type: object
    required: [subs_price, item_id]
    properties:
      subs_price:
        type: number
      status:
        type: boolean
        default: true
      item_id:
        type: integer
  SubscriptionResponse:
    type: object
    properties:
      success:
        type: boolean
      Message:
        type: string